Reorganize structure and add npm commands
